==> 06_RestAPI/01_RoadApi.ipynb <==
# 행정안전부 도로명주소 API
### 건물명으로부터 도로명주소 구하기

import requests
from urllib.parse import urlparse
import pandas as pd 

key_fd = open('roadapikey.txt', mode='r')
roadapi_key = key_fd.read(100)
key_fd.close()

road_url = 'https://www.juso.go.kr/addrlink/addrLinkApi.do'
url = f'{road_url}?confmKey={roadapi_key}&currentPage=1&countPerPage=10&keyword=강서구청&resultType=json'
urlparse(url).geturl()

results = requests.get(urlparse(url).geturl()).json()

addr_list = results['results']['juso']
len(addr_list)

results['results']['juso'][0]['roadAddr']

results['results']['juso'][0]

for addr in addr_list:
    print(addr['bdNm'], addr['roadAddr'])

### Exercise - 5개의 시청 및 구청의 도로명 주소 구하기

keyword_list = ['김포시청', '부천시청', '강서구청', '양천구청', '영등포구청']
road_url = 'https://www.juso.go.kr/addrlink/addrLinkApi.do'
road_addr_list = []
for keyword in keyword_list:
    url = f'{road_url}?confmKey={roadapi_key}&currentPage=1&countPerPage=10&keyword={keyword}&resultType=json'
    urlparse(url).geturl()
    results = requests.get(urlparse(url).geturl()).json()
    # 검색결과의 첫 번째의 도로명주소를 구한다
    addr = results['results']['juso'][0]['roadAddr']
    road_addr_list.append(addr)

road_addr_list

road_addr = pd.DataFrame({
    '키워드': keyword_list,
    '도로명주소': roadaddr_list
})
road_addr

road_addr.to_csv('../input/address.csv', sep=',', encoding='utf8', index=False)
