==> 06_RestAPI/02_kakaoAPI.ipynb <==
# 카카오맵 API
### 도로명주소로부터 위도, 경도 정보 구하기

import requests
from urllib.parse import urlparse
import pandas as pd 

key_fd = open('kakaopmapkey.txt', mode='r')
kmap_key = key_fd.read(100)
key_fd.close()

addr = '서울특별시 강서구 화곡로 302 (화곡동)'
url = 'https://dapi.kakao.com/v2/local/search/address.json?query=' + addr
result = requests.get(urlparse(url).geturl(), headers={"Authorization":"KakaoAK " + kmap_key}).json()
result

# 위도, 경도 추출
result['documents'][0]['y'], result['documents'][0]['x'] 

lat = float(result['documents'][0]['y'])
lng = float(result['documents'][0]['x'])
print(f'위도: {lat}   경도: {lng}')

### address.csv 파일 읽어서 위도, 경도 정보 추가하기

df = pd.read_csv('../input/address.csv')
df

df.index

lat_list = []; lng_list = []
for i in df.index:
    url = 'https://dapi.kakao.com/v2/local/search/address.json?query=' + df['도로명주소'][i]
    result = requests.get(urlparse(url).geturl(), headers={"Authorization":"KakaoAK " + kmap_key}).json()
    lat = float(result['documents'][0]['y'])
    lng = float(result['documents'][0]['x'])

    lat_list.append(lat); lng_list.append(lng)

df['lat'] = lat_list
df['lng'] = lng_list
df

df.to_csv('../input/address2.csv', sep=',', encoding='utf8', index=False)
