==> 06_RestAPI/04_folium.ipynb <==
# Folium 지도를 이용하여 표시하기

import pandas as pd 
import folium

folium.__version__

mapping = folium.Map(location=[37.550966, 126.849532], zoom_start=15)
mapping

mapping = folium.Map(location=[37.550966, 126.849532], tiles='Stamen Toner', zoom_start=13)
mapping

mapping = folium.Map(location=[37.550966, 126.849532], zoom_start=12)

folium.Marker(
    location=[37.517075, 126.866543],
    popup='양천구청',
    icon=folium.Icon(icon='cloud')
).add_to(mapping)

folium.Marker(
    location=[37.550966, 126.849532],
    popup='강서구청',
    icon=folium.Icon(color='green')
).add_to(mapping)

folium.Marker(
    location=[37.502277, 126.765004],
    popup='부천시청',
    icon=folium.Icon(color='red', icon='info-sign')
).add_to(mapping)
mapping

### address3.csv 파일을 읽어 지도 위에 표시하기

df = pd.read_csv('../input/address3.csv')
df

df['icon'] = ['map-marker', 'cloud', 'gift', 'info-sign', 'ok-circle']
df

# 날씨에 따른 마커의 색깔
color_dict = {'Clear': 'blue', 'Snow': 'white', 'Rain': 'gray', 'Extreme': 'red', 'Clouds': 'orange', 'Mist': 'green'}

mapping = folium.Map(location=[df.lat.mean(), df.lng.mean()], zoom_start=12)
mapping

mapping = folium.Map(location=[df.lat.mean(), df.lng.mean()], zoom_start=12)
for i in df.index:
    folium.Marker(
        location=[df.lat[i], df.lng[i]],
        popup=df['키워드'][i], 
        tooltip=f'{df.desc[i]}, {df.temp[i]}',
        icon=folium.Icon(color=color_dict[df.weather[i]], icon=df.icon[i])
    ).add_to(mapping)
mapping
